+++ src/main/kotlin/me/moeszyslak/taboo/commands/GuildConfigurationCommands.kt
@@ -10,12 +10,11 @@ import me.moeszyslak.taboo.extensions.requiredPermissionLevel
 import me.moeszyslak.taboo.services.Permission
 
 fun guildConfigurationCommands(configuration: Configuration) = commands("GuildConfiguration") {
-
     guildCommand("Setup") {
         description = "Setup a guild to use Taboo"
         requiredPermissionLevel = Permission.GUILD_OWNER
         execute {
-            if (configuration.hasGuildConfig(guild.id.longValue)) {
+            if (configuration.hasGuildConfig(guild.id)) {
                 respond("Guild configuration already exists. You can use commands to modify the config")
                 return@execute
             }
@@ -28,13 +27,12 @@ fun guildConfigurationCommands(configuration: Configuration) = commands("GuildCo
         }
     }
 
-
     guildCommand("Prefix") {
         description = "Set the prefix required for the bot to register a command."
         requiredPermissionLevel = Permission.STAFF
         execute(AnyArg("Prefix")) {
             val prefix = args.first
-            val config = configuration[guild.id.longValue] ?: return@execute
+            val config = configuration[guild.id] ?: return@execute
 
             config.prefix = prefix
             configuration.save()
@@ -48,9 +46,9 @@ fun guildConfigurationCommands(configuration: Configuration) = commands("GuildCo
         requiredPermissionLevel = Permission.STAFF
         execute(RoleArg) {
             val requiredRole = args.first
-            val config = configuration[(guild.id.longValue)] ?: return@execute
+            val config = configuration[guild.id] ?: return@execute
 
-            config.staffRole = requiredRole.id.longValue
+            config.staffRole = requiredRole.id
             configuration.save()
 
             respond("Required role set to ${requiredRole.name}")
@@ -62,9 +60,9 @@ fun guildConfigurationCommands(configuration: Configuration) = commands("GuildCo
         requiredPermissionLevel = Permission.STAFF
         execute(ChannelArg) {
             val logChannel = args.first
-            val config = configuration[(guild.id.longValue)] ?: return@execute
+            val config = configuration[guild.id] ?: return@execute
 
-            config.logChannel = logChannel.id.longValue
+            config.logChannel = logChannel.id
             configuration.save()
 
             respond("Logging channel set to ${logChannel.name}")
